moved the methods in the service and made my code DRY
